+++ drivers/sqlboiler-psql/driver/psql.go
@@ -399,8 +399,10 @@ func (p *PostgresDriver) TranslateColumnType(c drivers.Column) drivers.Column {
 			c.Type = "null.Int"
 		case "smallint", "smallserial":
 			c.Type = "null.Int16"
-		case "decimal", "numeric", "double precision":
+		case "decimal", "numeric":
 			c.Type = "types.NullDecimal"
+		case "double precision":
+			c.Type = "null.Float64"
 		case "real":
 			c.Type = "null.Float32"
 		case "bit", "interval", "bit varying", "character", "money", "character varying", "cidr", "inet", "macaddr", "text", "uuid", "xml":
@@ -458,8 +460,10 @@ func (p *PostgresDriver) TranslateColumnType(c drivers.Column) drivers.Column {
 			c.Type = "int"
 		case "smallint", "smallserial":
 			c.Type = "int16"
-		case "decimal", "numeric", "double precision":
+		case "decimal", "numeric":
 			c.Type = "types.Decimal"
+		case "double precision":
+			c.Type = "float64"
 		case "real":
 			c.Type = "float32"
 		case "bit", "interval", "uuint", "bit varying", "character", "money", "character varying", "cidr", "inet", "macaddr", "text", "uuid", "xml":
@@ -522,8 +526,10 @@ func getArrayType(c drivers.Column) string {
 		return "types.StringArray"
 	case "boolean":
 		return "types.BoolArray"
-	case "decimal", "numeric", "double precision", "real":
+	case "decimal", "numeric":
 		return "types.DecimalArray"
+	case "double precision", "real":
+		return "types.Float64Array"
 	default:
 		return "types.StringArray"
 	}